+++ Runtime/IGU/IGUObjects/IGUButton.cs
@@ -25,7 +25,7 @@ namespace Cobilas.Unity.Graphics.IGU.Elements {
             buttonStyle = GetDefaultValue(buttonStyle, GUI.skin.button);
             Rect rectTemp = new Rect(GetPosition(), myRect.Size);
 
-            Event current = Event.current;
+            Event current = IGUDrawer.IGUEvent;
 
             if (GUI.Button(rectTemp, GetGUIContent(DefaultContentIGUButton), buttonStyle)) {
                 if (IGUDrawer.Drawer.GetMouseButtonUp(myConfg.MouseType)) {